+++ lexicmap/cmd/kv/kv-reader.go
@@ -315,6 +315,10 @@ func (rdr *Reader) ReadDataOfAMaskAsList() ([]uint64, error) {
 	}
 	nKmers := int(be.Uint64(buf8))
 
+	if nKmers == 0 { // this hapens when no captured k-mer for a mask
+		return make([]uint64, 0), nil
+	}
+
 	// A list of k-mer and value pairs are intermittently saved in a []uint64
 	m := make([]uint64, 0, nKmers<<1)
 	// multiping 2.2 is because that some k-mers would have more than one locations,