+++ lexicmap/cmd/kv/kv-searcher.go
@@ -166,6 +166,10 @@ func (scr *Searcher) Search(kmers []uint64, p uint8, m int) (*[]*SearchResult, e
 	ttt := (uint64(1) << (k << 1)) - 1
 
 	for iQ, index := range scr.Indexes {
+		if len(index) == 0 { // this hapens when no captured k-mer for a mask
+			continue
+		}
+
 		// scope to search
 		// e.g., For a query ACGAC and p=3,
 		// kmers shared >=3 prefix are: ACGAA ... ACGTT.